Simplifying the ondisk structure of the repos
Removing a lot of the compexity in defining a repo and then its structure on disk.

Now, the definition of a repo is just the type (yum, apt) and source.  The name of the repo is the key name in the repositories hash in the config yaml.  It will be for the user to name it appropriately.  There are more and more 3rd party repos that are arch agnostic these days that forcing the declaration of the arch seemed excessive.

as a result the structure on disk is now:
/repo_base_location/<type>/<repo_name>

e.g.
/srv/repos/yum/rhel6_x86-64_updates