Fix webui handling brand new repos yet to be synced
Depending on the filesystem permissions and how the webserver is set up. The webapp may not be able to create the seed directories for new repos before the sync tool has been run and would create an error.

I will have to refactor a touch to definitivly separate the creation/syncing functions from the display functions.