New CLI Client
The sync-repo script has been replaced with a client to the APIs that is much more feature rich.  It also avoids confusion where repo files and logs are created by processes running as a user have permissions that mean that the web service can't use them.  Now everything is done in by the webservice.