Yum Sync Fixes
I had previously tried to force a repo synced with reposync for Yum repos to look like repos synced by http or local ones in terms of directory structure (i.e. repo/repodata, repo/*.rpm).  This was based on the assumption that the dir structure created was the result of reposync, and not the source repo.  It turns out that it is the source repo that defines the resulting directory structure and therefore not consistant.  So now, I just let it create whatever it wants and then let the createrepo program recurse the directories.  This does not change the url that would be used to access the repo in a yum config.